Fix Wsign-compare warnings

+++ src/libprojectM/projectM.cpp
@@ -900,7 +900,7 @@ void projectM::selectPrevious(const bool hardCut) {
             renderer->m_activePresetID--;
             selectPresetByName(renderer->m_presetList[renderer->m_activePresetID-1].name,true);
         }
-    } else if (settings().shuffleEnabled && presetHistory.size() >= 1 && presetHistory.back() != m_presetLoader->size() && !renderer->showmenu) { // if randomly browsing presets, "previous" should return to last random preset not the index--. Avoid returning to size() because that's the idle:// preset.
+    } else if (settings().shuffleEnabled && presetHistory.size() >= 1 && static_cast<std::size_t>(presetHistory.back()) != m_presetLoader->size() && !renderer->showmenu) { // if randomly browsing presets, "previous" should return to last random preset not the index--. Avoid returning to size() because that's the idle:// preset.
         presetFuture.push_back(m_presetPos->lastIndex());
         selectPreset(presetHistory.back());
         presetHistory.pop_back();
@@ -922,7 +922,7 @@ void projectM::selectNext(const bool hardCut) {
     if (isTextInputActive() && renderer->m_presetList.size() >= 1) // if search is active and there are search results
     {
         // if search menu is down, next is based on search terms.
-        if (renderer->m_activePresetID >= renderer->m_presetList.size()) {
+        if (static_cast<std::size_t>(renderer->m_activePresetID) >= renderer->m_presetList.size()) {
             // loop to top of page is at bottom
             renderer->m_activePresetID = 1;
             selectPresetByName(renderer->m_presetList[0].name,true);
@@ -931,7 +931,7 @@ void projectM::selectNext(const bool hardCut) {
             // otherwise move forward 
             selectPresetByName(renderer->m_presetList[renderer->m_activePresetID].name,true);
         }
-    } else if (settings().shuffleEnabled && presetFuture.size() >= 1 && presetFuture.front() != m_presetLoader->size() && !renderer->showmenu) { // if shuffling and we have future presets already stashed then let's go forward rather than truely move randomly.
+    } else if (settings().shuffleEnabled && presetFuture.size() >= 1 && static_cast<std::size_t>(presetFuture.front()) != m_presetLoader->size() && !renderer->showmenu) { // if shuffling and we have future presets already stashed then let's go forward rather than truely move randomly.
         presetHistory.push_back(m_presetPos->lastIndex());
         selectPreset(presetFuture.back());
         presetFuture.pop_back();
